Update sentiment data type to be double.

+++ portfolio/src/main/java/com/google/sps/data/Comment.java
@@ -18,7 +18,7 @@ public abstract class Comment {
 
   abstract String imageUrl();
 
-  abstract float sentiment();
+  abstract double sentiment();
 
   /** 
    * Returns a builder instance that can be used to create Comments.
@@ -35,7 +35,7 @@ public abstract class Comment {
     public abstract Builder setMood(String mood);
     public abstract Builder setNickname(String nickname);
     public abstract Builder setImageUrl(String imageUrl);
-    public abstract Builder setSentiment(float sentiment);
+    public abstract Builder setSentiment(double sentiment);
     public abstract Comment build();
   }
 }
